+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - An email is sent via GOV.UK Notify to all team leads when a project is
   created.
 - The Handover/new project form collects a trust SharePoint link.
+- The trust SharePoint link is shown on the project information.
 
 #### Changed